backend/pyproject.toml
[tool.poetry]
name = "memoryengine-backend"
version = "0.1.0"
description = "AI-powered knowledge management system backend"
authors = ["MemoryEngine Team"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
sqlalchemy = "^2.0.25"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.30.0"
greenlet = "^3.2.4"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
google-generativeai = "^0.3.2"
alembic = "^1.13.1"
pydantic = {extras = ["email"], version = "^2.5.3"}
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
httpx = "^0.26.0"
hypothesis = "^6.96.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
